for delete spot edit checked exceptions, added global exception, remove location when there are no more spots on it

+++ backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/service/impl/SimpleSpotService.java
@@ -1,12 +1,11 @@
 package at.ac.tuwien.sepm.groupphase.backend.service.impl;
 import at.ac.tuwien.sepm.groupphase.backend.entity.Spot;
-import at.ac.tuwien.sepm.groupphase.backend.exception.NotFoundException;
-import at.ac.tuwien.sepm.groupphase.backend.exception.NotFoundInDatabaseException;
+import at.ac.tuwien.sepm.groupphase.backend.repository.LocationRepository;
 import at.ac.tuwien.sepm.groupphase.backend.repository.SpotRepository;
 import at.ac.tuwien.sepm.groupphase.backend.service.SpotService;
+import org.hibernate.service.spi.ServiceException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.lang.invoke.MethodHandles;
@@ -16,9 +15,11 @@ public class SimpleSpotService implements SpotService {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
     private final SpotRepository spotRepository;
+    private final LocationRepository locationRepository;
 
-    public SimpleSpotService(SpotRepository spotRepository) {
+    public SimpleSpotService(SpotRepository spotRepository, LocationRepository locationRepository) {
         this.spotRepository = spotRepository;
+        this.locationRepository = locationRepository;
     }
 
     @Override
@@ -28,12 +29,15 @@ public class SimpleSpotService implements SpotService {
     }
 
     @Override
-    public void deleteById(Long id) throws NotFoundInDatabaseException {
+    public void deleteById(Long id) throws ServiceException {
         LOGGER.debug("Delete Spot with id {}", id);
-        try {
-            spotRepository.deleteById(id);
-        }catch (EmptyResultDataAccessException e){
-            throw new NotFoundInDatabaseException("Spot does not exist");
+        var spot = spotRepository.findById(id);
+        if(spot.isEmpty()){
+            throw new ServiceException("Spot does not exist");
+        }
+        spotRepository.deleteById(id);
+        if (spotRepository.findLocationWithSpot(spot.get().getLocation().getId()).isEmpty()){
+            locationRepository.deleteById(spot.get().getLocation().getId());
         }
     }
 }
